backend/requirements.txt
fastapi==0.115.12
uvicorn[standard]>=0.22.0,<0.24.0
pydantic>=2.2.0
python-dotenv>=1.0.0
langchain-openai>=0.0.1
langgraph>=0.1.0
supabase>=2.0.0
psycopg2-binary>=2.9.5
python-multipart>=0.0.5
jinja2>=3.1.2 